move tests that require Active Record.
This makes the `test/mass_assignment_security_test.rb` executable again.
